set up delete workshop function
