+++ src/App.css
@@ -24,6 +24,7 @@ body {
 
 a,
 a:visited {
+  text-decoration: none;
   color: #00d8ff
 }
