+++ src/pages/WorkshopDetails/WorkshopDetails.jsx
@@ -8,7 +8,7 @@ import * as workshopService from '../../services/workshopService'
 //css
 import styles from './WorkshopDetails.module.css'
 
-const WorkshopDetails = ({user}, {handleDeleteWorkshop}) => {
+const WorkshopDetails = ({user, handleDeleteWorkshop}) => {
   const { workshopId } = useParams()
   const [workshop, setWorkshop] = useState(null)
 
@@ -29,7 +29,7 @@ const WorkshopDetails = ({user}, {handleDeleteWorkshop}) => {
       <h1 className={styles.title}> {workshop.title} </h1>
       <h2> Hosted by {workshop.mentorInfo.name}</h2>
       <h3 className={styles.row}>
-        <div>Venue: {workshop.location.venueTitle} </div>
+        {/* <div>Venue: {workshop.location.venueTitle} </div> */}
         <div>Price: ${workshop.pricePerPerson} </div>
       </h3>
       <h3 className={styles.row}>
